+++ codec/common/cpu_core.h
