+++ codec/common/typedefs.h
@@ -37,7 +37,7 @@
 #include <limits.h>
 
 ////////////////////////////////////////////////////////////////////////////
-// NOTICE : ALL internal implement MUST use the pData type defined as below
+// NOTICE : ALL internal implement MUST use the data type defined as below
 //          ONLY except with the interface file !!!!!
 ////////////////////////////////////////////////////////////////////////////
 
@@ -62,7 +62,6 @@ typedef unsigned __int64 uint64_t;
 // FIXME:     all string type should be declared explicit as char.
 typedef char      str_t;
 typedef float     real32_t;
-
 #ifdef EPSN
 #undef EPSN
 #endif//EPSN
@@ -71,7 +70,6 @@ typedef float     real32_t;
 #ifndef NULL
 #define NULL 0
 #endif
-
 
 typedef bool bool_t;
 typedef int32_t BOOL_T;
@@ -84,5 +82,9 @@ typedef int32_t BOOL_T;
 #define TRUE    ((int32_t)1)
 #endif//TRUE
 
+#ifndef void_t
+#define void_t void
+#endif
 
 #endif //WELS_TYPE_DEFINES_H__
+
