+++ codec/console/dec/src/h264dec.cpp
@@ -45,7 +45,7 @@
 #include "codec_app_def.h"
 #include "codec_api.h"
 #include "read_config.h"
-#include "../../decoder/core/inc/typedefs.h"
+#include "typedefs.h"
 #include "../../decoder/core/inc/measure_time.h"
 #include "d3d9_utils.h"
 #include "logging.h"
